EPMRPP-64796: code review fixes - 3
