+++ .eslintrc
@@ -31,6 +31,7 @@
     "import/prefer-default-export": 0,
     "@typescript-eslint/ban-ts-ignore": 0,
     "@typescript-eslint/no-implied-eval": 0,
-    "dot-notation": 0
+    "dot-notation": 0,
+    "@typescript-eslint/naming-convention": 0
   }
 }
