--- examples/rp.json
@@ -1,16 +1,0 @@
-{
-  "token": "00000000-0000-0000-0000-000000000000",
-  "endpoint": "https://your.reportportal.server/api/v1",
-  "project": "Your project",
-  "launch": "Playwright test",
-  "attributes": [
-    {
-      "key": "key",
-      "value": "value"
-    },
-    {
-      "value": "value"
-    }
-  ],
-  "description": "Your launch description"
-}
