+++ src/__tests__/reporter/startSuiteTestReporting.spec.ts
@@ -14,51 +14,66 @@
  *  limitations under the License.
  */
 
-import MyReporter from '../../reporter';
+import RPReporter from '../../reporter';
 import { mockConfig } from '../mocks/configMock';
-import { RPClientMock } from '../mocks/RPClientMock';
+import { mockedTestParams, RPClientMock } from '../mocks/RPClientMock';
 import { StartTestObjType } from '../../models';
 import { TEST_ITEM_TYPES } from '../../constants';
-import * as utils from '../../utils';
 
-describe('start report suite', () => {
-  jest.spyOn(utils, 'getConfig').mockImplementation(() => mockConfig);
-  const reporter = new MyReporter();
+describe('start reporting suite/test', () => {
+  const reporter = new RPReporter(mockConfig);
   reporter.client = new RPClientMock(mockConfig);
   reporter.launchId = 'tempLaunchId';
-  const testParams = {
-    title: 'test',
-    parent: {
-      title: 'suiteName',
-    },
-  };
-
-  const startSuiteObj: StartTestObjType = {
-    startTime: reporter.client.helpers.now(),
-    name: testParams.parent.title,
-    type: TEST_ITEM_TYPES.SUITE,
-  };
-
   const spyOnTestBegin = jest.spyOn(reporter, 'onTestBegin');
 
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
-  reporter.onTestBegin(testParams);
+  reporter.onTestBegin(mockedTestParams);
 
-  const expectedSuite = new Map([['suiteName', { id: 'tempTestItemId', name: 'suiteName' }]]);
+  describe('start suite report', () => {
+    const expectedSuite = new Map([['suiteName', { id: 'tempTestItemId', name: 'suiteName' }]]);
+    const startSuiteObj: StartTestObjType = {
+      startTime: reporter.client.helpers.now(),
+      name: mockedTestParams.parent.title,
+      type: TEST_ITEM_TYPES.SUITE,
+    };
 
-  test('client.startTestItem should be called with corresponding params', () => {
-    expect(spyOnTestBegin).toHaveBeenCalledTimes(1);
-    expect(reporter.client.startTestItem).toHaveBeenCalledWith(startSuiteObj, reporter.launchId);
+    test('client.startTestItem should be called with corresponding params', () => {
+      expect(spyOnTestBegin).toHaveBeenCalledTimes(1);
+      expect(reporter.client.startTestItem).toHaveBeenCalledWith(startSuiteObj, reporter.launchId);
+    });
+
+    test('reporter.suites should be updated', () => {
+      expect(reporter.suites).toEqual(expectedSuite);
+    });
   });
 
-  test('reporter.suites should be updated', () => {
-    expect(reporter.suites).toEqual(expectedSuite);
+  describe('start tests report', () => {
+    const expectedTestItems = new Map([['test', { id: 'tempTestItemId', name: 'test' }]]);
+    const parentId = 'tempTestItemId';
+    const startTestObj: StartTestObjType = {
+      startTime: reporter.client.helpers.now(),
+      name: mockedTestParams.title,
+      type: TEST_ITEM_TYPES.STEP,
+    };
+
+    test('client.startTestItem should be called with corresponding params', () => {
+      expect(spyOnTestBegin).toHaveBeenCalledTimes(1);
+      expect(reporter.client.startTestItem).toHaveBeenCalledWith(
+        startTestObj,
+        reporter.launchId,
+        parentId,
+      );
+    });
+
+    test('reporter.testItems should be updated', () => {
+      expect(reporter.testItems).toEqual(expectedTestItems);
+    });
   });
 });
 
 describe('suite in suite case', () => {
-  const reporter = new MyReporter();
+  const reporter = new RPReporter(mockConfig);
   reporter.client = new RPClientMock(mockConfig);
   reporter.launchId = 'tempLaunchId';
   const testParams = {
@@ -89,23 +104,5 @@ describe('suite in suite case', () => {
   ]);
   test('parent and child suites should be updated', () => {
     expect(reporter.suites).toEqual(expectedSuites);
-  });
-});
-
-describe('finish report suite', () => {
-  const reporter = new MyReporter();
-  reporter.client = new RPClientMock(mockConfig);
-  reporter.launchId = 'tempLaunchId';
-  reporter.suites = new Map([['suiteName', { id: 'tempTestItemId', name: 'suiteName' }]]);
-  reporter.onEnd();
-
-  test('client.finishTestItem should be called with suite id', () => {
-    expect(reporter.client.finishTestItem).toHaveBeenCalledWith('tempTestItemId', {
-      endTime: reporter.client.helpers.now(),
-    });
-  });
-
-  test('suites should be reset', () => {
-    expect(reporter.suites).toEqual(new Map());
   });
 });
